+++ src/api/mod.rs
@@ -1,8 +1,15 @@
 //! API helpers
 
 /// Empty struct for when no Spec is required
+///
+/// Not using [`()`](https://doc.rust-lang.org/stable/std/primitive.unit.html), because serde's
+/// [`Deserialize`](https://docs.rs/serde/1.0.104/serde/trait.Deserialize.html) `impl` is too strict.
 #[derive(Clone, Deserialize, Serialize, Default)]
-pub struct Void {}
+pub struct NotUsed {}
+
+/// Use [`NotUsed`](notused.struct.html) instead. Renamed to avoid confusion with [`void::Void`](https://docs.rs/void/1.0.2/void/enum.Void.html).
+#[deprecated]
+pub type Void = NotUsed;
 
 mod reflector;
 pub use self::reflector::Reflector;